Split 'atlas.py' into a Texture subpackage
