+++ py3dtilers/Common/__init__.py
@@ -1,7 +1,6 @@
 from .kd_tree import kd_tree
 from .object_to_tile import ObjectToTile, ObjectsToTile
 from .tree_with_children_and_parent import TreeWithChildrenAndParent
-from .atlas import getTexture, createTextureAtlas
 from .group import Groups
 from .polygon_extrusion import ExtrudedPolygon
 from .lod_node import LodNode, Lod1Node, LoaNode
@@ -18,6 +17,4 @@ __all__ = ['kd_tree',
            'LodNode',
            'Lod1Node',
            'LoaNode',
-           'LodTree',
-           'getTexture',
-           'createTextureAtlas']
+           'LodTree']
