+++ py3dtilers/Texture/__init__.py
@@ -1,0 +1,9 @@
+from .atlas_rectangle import Rectangle
+from .atlas_node import Node
+from .atlas import Atlas
+from .texture import Texture
+
+__all__ = ['Rectangle',
+           'Node',
+           'Atlas',
+           'Texture']
