+++ py3dtilers/Texture/atlas_node.py
@@ -1,66 +1,6 @@
 import numpy as np
-from io import BytesIO
 from PIL import Image
-
-# This file implement the solution described here
-# https://blackpawn.com/texts/lightmaps/
-# to properly pack multiple texture in a Atlas by creating a tree
-# of rectangle representing the Atlas.
-
-
-class Rectangle(object):
-    """
-    The class that represents a rectangle in the atlas by its position, width
-    and height.
-    """
-
-    def __init__(self, left, top, right, bottom):
-        self.left = left
-        self.right = right
-        self.top = top
-        self.bottom = bottom
-        self.width = right - left
-        self.height = bottom - top
-
-    def setSize(self, newWidth, newHeight):
-        self.width = newWidth
-        self.height = newHeight
-
-    def get_top(self):
-        return self.top
-
-    def get_bottom(self):
-        return self.bottom
-
-    def get_right(self):
-        return self.right
-
-    def get_left(self):
-        return self.left
-
-    def get_width(self):
-        return self.width
-
-    def get_height(self):
-        return self.height
-
-    def fits(self, img):
-        """
-        :param img: A pillow image
-        :rtype boolean: Whether the image fits in the rectangle or no
-                        i.e if the image is smaller than the rectangle
-        """
-        imageWidth, imageHeight = img.size
-        return imageWidth <= self.width and imageHeight <= self.height
-
-    def perfect_fits(self, img):
-        """
-        :param img: A pillow image
-        :rtype boolean: Whether the image prefectly fits in the rectangle or no,
-                    i.e if the image have the exact same size of the rectangle
-        """
-        imageWidth, imageHeight = img.size
-        return imageWidth == self.width and imageHeight == self.height
+from ..Texture import Rectangle
 
 
 class Node(object):
@@ -233,149 +173,3 @@ class Node(object):
                 # warning : in order to be written correctly, the GLTF writter
                 # expects data to be in float32
                 uvs[i][y] = np.array([new_u, new_v], dtype=np.float32)
-
-
-def updateUvs(uvs, rect):
-    offsetX = rect[0]
-    offsetY = rect[1]
-    if rect[2] != rect[0]:
-        ratioX = 1 / (rect[2] - rect[0])
-    else:
-        ratioX = 1
-    if rect[3] != rect[1]:
-        ratioY = 1 / (rect[3] - rect[1])
-    else:
-        ratioY = 1
-
-    for i in range(0, len(uvs)):
-        for y in range(0, 3):
-            new_u = (uvs[i][y][0] - offsetX) * ratioX
-            new_v = (uvs[i][y][1] - offsetY) * ratioY
-            # warning : in order to be written correctly, the GLTF writter
-            # expects data to be in float32
-            uvs[i][y] = np.array([new_u, new_v], dtype=np.float32)
-
-
-def computeArea(size):
-    """
-    :param size : an array with a width and a height of a texture
-    :rtype float: the area of the texture
-    """
-    width, height = size
-    return width * height
-
-
-def getTexture(textureUri, objects_type, cursor, triangles):
-    """
-    :param textureUri : a texture Uri in the database
-    :param objects_type: a class name among CityMCityObject derived classes.
-                    For example, objects_type can be "CityMBuilding".
-    :param cursor: a database access cursor
-    :rtype pillow.image:
-    """
-    imageBinaryData = objects_type.retrieve_textures(
-        cursor,
-        textureUri,
-        objects_type)
-    LEFT_THUMB = imageBinaryData[0][0]
-    stream = BytesIO(LEFT_THUMB)
-    image = Image.open(stream).convert("RGBA")
-
-    return cropImage(image, triangles)
-
-
-def cropImage(image, triangles):
-    minX = 2
-    maxX = -1
-
-    minY = 2
-    maxY = -1
-
-    texture_size = image.size
-    for uv_triangle in triangles:
-        for uv in uv_triangle:
-            if uv[0] < minX:
-                minX = uv[0]
-            if uv[0] > maxX:
-                maxX = uv[0]
-            if uv[1] < minY:
-                minY = uv[1]
-            if uv[1] > maxY:
-                maxY = uv[1]
-
-    cropped_image = image.crop((minX * texture_size[0], minY * texture_size[1], maxX * texture_size[0], maxY * texture_size[1]))
-
-    updateUvs(triangles, [minX, minY, maxX, maxY])
-    return cropped_image
-
-
-def multipleOf2(nb):
-    """
-    :param nb: a number
-    :rtype float: The first multiple of 2 greater than the number
-    """
-    i = 1
-    while i < nb:
-        i *= 2
-    return i
-
-
-def computeAtlasTree(textures_sorted):
-    """
-    :param textures_sorted:  A dictionnary, with building_id as key,
-                        and pillow image as value.
-    :rtype node: the root node of the atlas tree
-    """
-    surfaceAtlas = 0
-    for key, image in textures_sorted:
-        # Add the surface of the current texture to the atlas one
-        surfaceAtlas += computeArea(image.size)
-
-    # We estimate the size of the atlas to be around the nearest power of 2
-    # of the squareroot of all surfaces combined. There will be cases
-    # where this does not work, since this creates a square.
-    sizeOfAtlas = multipleOf2(np.sqrt(surfaceAtlas))
-
-    rect = Rectangle(0, 0, (sizeOfAtlas // 2) + 1, (sizeOfAtlas // 2) + 1)
-    node_root = None
-    it = 0
-    while node_root is None:
-        node_root = Node(rect)
-        axis = it % 2
-        axisX = 1 - axis
-        axisY = 1 - (1 - axis)
-        for key, image in textures_sorted:
-            node_root = node_root.insert(image, key)
-            if node_root is None:
-                # If the node_root is None, this means that the estimation
-                # of the size of the atlas is wrong and must be enlarged
-                rect = Rectangle(
-                    0,
-                    0,
-                    int(rect.get_width() + rect.get_width() * 0.5 * axisX),
-                    int(rect.get_height() + rect.get_height() * 0.5 * axisY)
-                )
-                it += 1
-                break
-    node_root.set_tile_number()
-    return node_root
-
-
-def createTextureAtlas(objects):
-    objects_with_id_key = dict()
-    textures_with_id_key = dict()
-
-    for object_to_tile in objects:
-        objects_with_id_key[object_to_tile.get_id()] = object_to_tile.geom
-        textures_with_id_key[object_to_tile.get_id()] = object_to_tile.get_texture()
-
-    # Sort textures by size, starting by the biggest one
-    textures_sorted = sorted(textures_with_id_key.items(),
-                             key=lambda t: computeArea(t[1].size), reverse=True)
-
-    atlasTree = computeAtlasTree(textures_sorted)
-    tile_number = atlasTree.get_tile_number()
-
-    atlasTree.createAtlasImage(objects_with_id_key, tile_number)
-
-    return tile_number
